feat : 주문 조회 수량 누락 수정

+++ src/main/java/store/onuljang/controller/response/AdminReservationListResponse.java
@@ -18,7 +18,7 @@ public record AdminReservationListResponse(
         LocalDate orderDate,
         String productName,
         String userName,
-        long quantity,
+        int quantity,
         BigDecimal amount,
         ReservationStatus status)
     {
